LPS-134356 [do not merge] get active locales from the context display object for web content. That object already had the right method, it just was private. I wrapped it with a public one which turns the array into a List<Srring> so that we can pass it to the tag. This illustrates the kind of job that apps should do in order to integrate with the translation admin.
With this commit, only translated languages will be used as active ones, meaning that if a web content is translated to 3 languages, those 3 will be shown in the selector. If we add another one via translation admin, and provide translation for it in some field (say, the "Title"), then next time we edit it the selector will show 4 languages. Note that if no translation are provided though, then the system won't persist it so next time we edit the content only 3 languages will be in the selector.

Finally, note that this integration lacks a crucial piece of logic. The journal-web module is not listening to the language deletion event via activelocales Atom state subscription, meaning that it can't delete the content written in the web content fields (such as the one named "content" in the basic web content). Therefore, if one inputs something in that field for a language, and then deletes the language via translation admin modal, and then save, the language is not totally removed: only localized-input values will have been removed (title, description and friendly url), but not content. As a result the language is not effectively removed and next time we edit the content, we'll see the language in the selector in "UNTRANSLATED" state. That's a side effect from the fact that app should do the job of removing the "content" field value for the removed language.