LPD-17128 Add basic webpack build to liferay-sample-theme-css-1. After theme CSS gradle build task, build:  * Cleans previous build artifacts in buildTheme folder, if any  * Hashes buildTheme generated artifacts main.css and clay.css  * Instructs client-extension.yaml to bundle hashed versions of main.css and clay.css
Now, ThemeCSS CX css resource names are hashed. As a result, build/liferay-client-extension-build/ needs to be clean whenever there is a change in the source css files, otherwise, old hashed files will remain there. This is to be fixed in LPD-19346. In the meantime, this CX needs gw clean deploy